HPCC-8410 Move daregress to CPPUNIT
Moving daregress to CPPUNIT and adding more tests that were not commonly used.
Also making the CDistributedFile::detachLogical() private, to avoid misuse and
moving some common CPPUNIT code to a shared header.
