+++ dali/CMakeLists.txt
@@ -26,7 +26,6 @@ HPCC_ADD_SUBDIRECTORY (dfuplus "CLIENTTOOLS")
 HPCC_ADD_SUBDIRECTORY (dfuxref "PLATFORM")
 HPCC_ADD_SUBDIRECTORY (dfuXRefLib "PLATFORM")
 HPCC_ADD_SUBDIRECTORY (ft)
-HPCC_ADD_SUBDIRECTORY (regress "PLATFORM")
 HPCC_ADD_SUBDIRECTORY (daliadmin "PLATFORM")
 HPCC_ADD_SUBDIRECTORY (server "PLATFORM")
 HPCC_ADD_SUBDIRECTORY (treeview "PLATFORM")
