+++ roxie/ccd/ccdkey.cpp
@@ -1963,8 +1963,7 @@ MODULE_EXIT()
 //=======================================================================================================
 
 #ifdef _USE_CPPUNIT
-#include <cppunit/extensions/HelperMacros.h>
-#define ASSERT(a) { if (!(a)) CPPUNIT_ASSERT(a); }
+#include "unittests.hpp"
 
 class InMemoryIndexTest : public CppUnit::TestFixture  
 {
