+++ roxie/roxiemem/roxiemem.cpp
@@ -3269,8 +3269,7 @@ extern void setDataAlignmentSize(unsigned size)
 
 //============================================================================================================
 #ifdef _USE_CPPUNIT
-#include <cppunit/extensions/HelperMacros.h>
-#define ASSERT(a) { if (!(a)) CPPUNIT_ASSERT(a); }
+#include "unittests.hpp"
 
 namespace roxiemem {
