+++ roxie/roxiemem/roxierow.cpp
@@ -495,8 +495,7 @@ IRowAllocatorMetaActIdCache *createRowAllocatorCache(IRowAllocatorMetaActIdCache
 
 
 #ifdef _USE_CPPUNIT
-#include <cppunit/extensions/HelperMacros.h>
-#define ASSERT(a) { if (!(a)) CPPUNIT_ASSERT(a); }
+#include "unittests.hpp"
 
 namespace roxiemem {
