+++ roxie/roxiemem/roxierowbuff.cpp
@@ -219,8 +219,7 @@ void RoxieSimpleInputRowArray::transferFrom(RoxieOutputRowArray & donor)
 //============================================================================================================
 
 #ifdef _USE_CPPUNIT
-#include <cppunit/extensions/HelperMacros.h>
-#define ASSERT(a) { if (!(a)) CPPUNIT_ASSERT(a); }
+#include "unittests.hpp"
 
 namespace roxiemem {
